+++ priv/repo/migrations/20260417124227_add_enrolled_child_names_to_conversation_summaries.exs
@@ -1,0 +1,9 @@
+defmodule KlassHero.Repo.Migrations.AddEnrolledChildNamesToConversationSummaries do
+  use Ecto.Migration
+
+  def change do
+    alter table(:conversation_summaries) do
+      add :enrolled_child_names, {:array, :string}, default: []
+    end
+  end
+end